test(dbt-cloud): apply `pytest` markers to split test execution
